Amélioration du composant weather-forecast
